+++ MiscMapActionsProperties/Framework/Location/Panorama.cs
@@ -440,9 +440,11 @@ internal sealed class PanoramaBackground(GameLocation location) : Background(loc
 /// </summary>
 internal static class Panorama
 {
-    internal static readonly string MapProp_PanoramaPrefix = $"{ModEntry.ModId}_Panorama";
+    internal const string MapProp_PanoramaPrefix = $"{ModEntry.ModId}_Panorama";
 
-    internal static readonly string Asset_Panorama = $"{ModEntry.ModId}/Panorama";
+    internal static readonly IAssetName Asset_Panorama = ModEntry.help.GameContent.ParseAssetName(
+        $"{ModEntry.ModId}/Panorama"
+    );
 
     internal static void Register()
     {
@@ -474,7 +476,7 @@ internal static class Panorama
 
     private static Dictionary<string, PanoramaData>? _bgData = null;
     internal static Dictionary<string, PanoramaData> BgData =>
-        _bgData ??= Game1.content.Load<Dictionary<string, PanoramaData>>(Asset_Panorama);
+        _bgData ??= ModEntry.help.GameContent.Load<Dictionary<string, PanoramaData>>(Asset_Panorama);
 
     private static void OnAssetRequested(object? sender, AssetRequestedEventArgs e)
     {